COMP9323_Project/OLC/backend/requirements.txt
arrow==1.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==2.5.1
Flask==2.0.1
oss2==2.15.0
PyMySQL==1.0.2
pytest==6.2.4
requests==2.26.0
setuptools==57.0.0
Werkzeug==2.0.1
wheel==0.36.2
